requirements.txt
robotframework==7.2.2
robotframework-requests==0.9.7
robotframework-seleniumlibrary==6.7.1
robotframework-pabot==4.2.0
python-dotenv==1.1.0
requests==2.32.3
pytest==8.3.5
pytest-html==4.1.1